requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pdfplumber
groq
python-dotenv